orientation-event/motion/page-visibility.https.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/page-visibility/resources/window_state_context.js"></script>
<script src="../resources/orientation-event-helpers.js"></script>
<script>
'use strict';

promise_test(async (t, sensorProvider) => {
  await setMotionSensorsPermissions();
  await createVirtualSensors([['accelerometer', true],['linear-acceleration', true],['gyroscope', true]]);
  t.add_cleanup(async () => {
    await removeVirtualSensors(['accelerometer', 'linear-acceleration', 'gyroscope']);
  });
  startFetchingEventData(t, 'devicemotion');

  const motionData1 = generateMotionData(1, 2, 3,
                                         4, 5, 6,
                                         7, 8, 9);

  const motionData2 = generateMotionData(11, 12, 13,
                                         14, 15, 16,
                                         17, 18, 19);

  await setMockMotionData(t, motionData1, true, true, true);
  const event = getExpectedMotionEvent(motionData1);
  await waitForEvent(event);

  const {minimize, restore} = window_state_context(t);
  await minimize();
  assert_true(document.hidden);

  let hiddenEventPromise = new Promise((resolve, reject) => {
    window.addEventListener(
      'devicemotion',
      event => {
        if (document.hidden) {
          reject();
        } else {
          resolve();
        }
      },
      { once: true });
  });

  // devicemotion events fire every 16ms when active, so we sleep for a while
  // here to make sure we're not actually getting any events.
  await new Promise(resolve => { t.step_timeout(resolve, 100); });
  await restore();
  assert_false(document.hidden);

  // Restart suspended sensors caused by minimizing browser window.
  await setMockMotionData(t, motionData2, true, true, true);
  const event2 = getExpectedMotionEvent(motionData2);

  return Promise.all([hiddenEventPromise, waitForEvent(event2)]);
}, 'Tests to check that devicemotion events are not fired when the page is not visible.');
</script>
</body>
</html>
